core/sync: extract separate push
